+++ src/test/java/io/json/compare/matcher/diffs/JsonArrayDiffTests.java
@@ -103,11 +103,13 @@ public class JsonArrayDiffTests {
 
     @Test
     public void compareJsonArraysAndCheckForMultipleDoNotMatchAnyDifferences() {
-        String expected = "[{\"lorem\":\"ipsum\"},\"!.*\",\"1\",\"!.*\"]";
+        String expected = "[{\"lorem\":\"ipsum\"},\"!.*\",\"1\",\"!.*\",\"!.*\"]";
         String actual = "[\"a\",true,1,{\"lorem\":\"ipsum\"}]";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 1 DIFFERENCE.*" +
-                "Expected condition \"\\Q!.*\\E\" from position 2 was not met. Actual JSON ARRAY has extra elements.*"));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
+                "Expected condition \"\\Q!.*\\E\" from position 2 was not met. Actual JSON ARRAY has extra elements.*" +
+                "Expected condition \"\\Q!.*\\E\" from position 4 was not met. Actual JSON ARRAY has extra elements.*" +
+                "Expected condition \"\\Q!.*\\E\" from position 5 was not met. Actual JSON ARRAY has extra elements.*"));
         JSONCompare.assertNotMatches(expected, actual);
     }
 
@@ -116,8 +118,9 @@ public class JsonArrayDiffTests {
         String expected = "[{\"lorem\":\"ipsum\"},\"!.*\",\"1\",{\"lorem2\":\"ipsum2\",\"lorem3\":\"ipsum3\"}]";
         String actual = "[\"a\",true,1,{\"lorem2\":\"ipsum-updated\"}]";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual, new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_STRICT_ORDER))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
                 "JSON ARRAY elements differ at position 1:.*lorem.*ipsum.*diffs.*Different JSON types: expected ObjectNode but got TextNode.*" +
+                "Expected condition \"\\Q!.*\\E\" from position 2 was not met. Actual JSON ARRAY has extra elements.*" +
                 "JSON ARRAY elements differ at position 4:.*lorem2.*ipsum2.*lorem3.*ipsum3.*diffs.*" +
                 "lorem2 ->.*Expected value: \"ipsum2\" But got: \"ipsum-updated\".*" +
                 "Field 'lorem3' was NOT FOUND.*"));
@@ -139,17 +142,19 @@ public class JsonArrayDiffTests {
         String expected = "[{\"lorem\":\"ipsum\"},\"c\",\"!1\",true]";
         String actual = "[\"a\",true,1,{\"lorem\":\"ipsum\"}]";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual, new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
                 "Expected element from position 2 was NOT FOUND.*\"c\".*" +
-                "Expected element from position 3 was FOUND.*\"!1\".*"));
+                "Expected element from position 3 was FOUND.*\"!1\".*" +
+                "Actual JSON ARRAY has extra elements.*"));
         JSONCompare.assertNotMatches(expected, actual, new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE)));
 
         String expected1 = "[{\"lorem\":\"ipsum\"},\"c\",\"!1\",true]";
         String actual1 = "[\"a\",true,1,{\"lorem\":\"ipsum\"},-10.02]";
         AssertionError error1 = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected1, actual1, new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE))));
-        assertTrue(error1.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+        assertTrue(error1.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
                 "Expected element from position 2 was NOT FOUND.*\"c\".*" +
-                "Expected element from position 3 was FOUND.*\"!1\".*"));
+                "Expected element from position 3 was FOUND.*\"!1\".*" +
+                "Actual JSON ARRAY has extra elements.*"));
         JSONCompare.assertNotMatches(expected1, actual1, new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE)));
     }
 
@@ -159,14 +164,15 @@ public class JsonArrayDiffTests {
         String actual = "[\"a\",true,1,{\"lorem\":\"ipsum\"},-10.02]";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual,
                 new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE, CompareMode.JSON_ARRAY_STRICT_ORDER))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 4 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 5 DIFFERENCE.*" +
                 "JSON ARRAY elements differ at position 1.*\"lorem\".*\"ipsum\".*diffs.*" +
                 "Different JSON types: expected ObjectNode but got TextNode.*" +
                 "JSON ARRAY elements differ at position 2.*" +
                 "\"c\".*diffs.*Expected value: \"c\" But got: true.*" +
                 "Expected element from position 3 was FOUND.*\"!1\".*" +
                 "JSON ARRAY elements differ at position 4.*true.*diffs.*" +
-                "Different JSON types: expected BooleanNode but got ObjectNode.*"));
+                "Different JSON types: expected BooleanNode but got ObjectNode.*" +
+                "Actual JSON ARRAY has extra elements.*"));
         JSONCompare.assertNotMatches(expected, actual,
                 new HashSet<>(Arrays.asList(CompareMode.JSON_ARRAY_NON_EXTENSIBLE, CompareMode.JSON_ARRAY_STRICT_ORDER)));
     }