+++ src/test/java/io/json/compare/matcher/diffs/JsonObjectDiffTests.java
@@ -99,8 +99,9 @@ public class JsonObjectDiffTests {
         String expected = "{\"b\":2,\"a\":1,\"c\":{\"c2\":\"lorem2\",\"!.*\":\".*\"},\"!.*\":\".*\"}";
         String actual = "{\"a\":1,\"b\":2,\"c\":{\"c1\":\"lorem1\",\"c2\":\"lorem2\"},\"d\":1}";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 1 DIFFERENCE.*" +
-                "c -> Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*"));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+                "c -> Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*" +
+                "Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*"));
         JSONCompare.assertNotMatches(expected, actual);
 
         String expected1 = "{\"b\":2,\"a\":1,\"c\":{\"c2\":\"lorem2\",\"!.*\":\".*\"},\"!.*\":\".*\"}";
@@ -123,8 +124,9 @@ public class JsonObjectDiffTests {
         String expected = "{\"a\":1,\"c\":{\"c1\":\"lorem2\",\".*\":\".*\"},\"!.*\":\".*\"}";
         String actual = "{\"a\":1,\"b\":2,\"c\":{\"c1\":\"lorem2\"}}";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 1 DIFFERENCE.*" +
-                "c -> Field \\Q'.*'\\E was NOT FOUND.*"));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+                "c -> Field \\Q'.*'\\E was NOT FOUND.*" +
+                "Expected condition \\Q'!.*'\\E was not met. Actual JSON OBJECT has extra fields.*"));
         JSONCompare.assertNotMatches(expected, actual);
     }
 
@@ -134,9 +136,10 @@ public class JsonObjectDiffTests {
         String actual = "{\"a\":1,\"b\":2,\"c\":{\"c2\":\"lorem2\",\"c1\":\"lorem1\"},\"x\":0}";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual,
                 new HashSet<>(Collections.singletonList(CompareMode.JSON_OBJECT_NON_EXTENSIBLE))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 2 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
                 "a ->.*Expected value: 0 But got: 1.*" +
-                "c -> Actual JSON OBJECT has extra fields.*"));
+                "c -> Actual JSON OBJECT has extra fields.*" +
+                "Actual JSON OBJECT has extra fields.*"));
         JSONCompare.assertNotMatches(expected, actual);
     }
 
@@ -146,20 +149,22 @@ public class JsonObjectDiffTests {
         String actual = "{\"a\":1, \"u\":[3,1,false], \"b\":2,\"c\":{\"c2\":\"lorem2\",\"c1\":\"lorem1\"},\"x\":0}";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual,
                 new HashSet<>(Collections.singletonList(CompareMode.JSON_OBJECT_NON_EXTENSIBLE))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 4 DIFFERENCE.*" +
                 "a ->.*Expected value: 0 But got: 1.*" +
                 "c -> Actual JSON OBJECT has extra fields.*" +
-                "u ->.*Expected condition \\Q\"!.*\"\\E from position 3 was not met. Actual JSON ARRAY has extra elements.*"));
+                "u ->.*Expected condition \\Q\"!.*\"\\E from position 3 was not met. Actual JSON ARRAY has extra elements.*" +
+                "Actual JSON OBJECT has extra fields.*"));
         JSONCompare.assertNotMatches(expected, actual);
 
         String expected1 = "{\"b\":2,\"a\":0,\"c\":{\"c1\":\"lorem1\", \"c2\":\"incorrect\"},\"u\":[3,1,\"!.*\"]}";
         String actual1 = "{\"a\":1, \"u\":[3,1,false], \"b\":2,\"c\":{\"c2\":\"lorem2\",\"c1\":\"lorem1\"},\"x\":0}";
         error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected1, actual1,
                 new HashSet<>(Collections.singletonList(CompareMode.JSON_OBJECT_NON_EXTENSIBLE))));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 4 DIFFERENCE.*" +
                 "a ->.*Expected value: 0 But got: 1.*" +
                 "c -> c2 ->.*Expected value: \"incorrect\" But got: \"lorem2\".*" +
-                "u ->.*Expected condition \\Q\"!.*\"\\E from position 3 was not met. Actual JSON ARRAY has extra elements.*"));
+                "u ->.*Expected condition \\Q\"!.*\"\\E from position 3 was not met. Actual JSON ARRAY has extra elements.*" +
+                "Actual JSON OBJECT has extra fields.*"));
         JSONCompare.assertNotMatches(expected, actual);
     }
 
@@ -288,13 +293,15 @@ public class JsonObjectDiffTests {
         String expected1 = "{\".*\":\"test\", \"records\":[1, \".*\", 3, \"!.*\"], \"otherRecords\":[4, \"!.*\"], \"!.*\":\".*\"}";
         String actual1 = "{\"names\":\"test1\", \"records\":[2,1,4,3], \"otherRecords\":[1,2], \"another\":\"record\"}";
         AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected1, actual1));
-        assertTrue(error.getMessage().matches("(?s).*FOUND 6 DIFFERENCE.*" +
+        assertTrue(error.getMessage().matches("(?s).*FOUND 8 DIFFERENCE.*" +
                 "\\Q.*\\E ->.*Expected value: \"test\" But got: \"test1\".*" +
                 "\\Q.*\\E ->.*Different JSON types: expected TextNode but got ArrayNode.*" +
                 "\\Q.*\\E ->.*Different JSON types: expected TextNode but got ArrayNode.*" +
                 "\\Q.*\\E ->.*Expected value: \"test\" But got: \"record\".*" +
                 "records ->.*Expected condition \"\\Q!.*\\E\" from position 4 was not met. Actual JSON ARRAY has extra elements.*" +
-                "otherRecords ->.*Expected element from position 1 was NOT FOUND.*4.*"));
+                "otherRecords ->.*Expected element from position 1 was NOT FOUND.*4.*" +
+                "otherRecords ->.*Expected condition \"\\Q!.*\\E\" from position 2 was not met. Actual JSON ARRAY has extra elements.*" +
+                "Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*"));
 
         String expected2 = "{\"name\":\"test\", \"records\":[1, \".*\", 3, 4, \".*\"], \"otherRecords\":[4, \"!.*\"], \".*\":\".*\"}";
         String actual2 = "{\"names\":\"test1\", \"records\":[2,1,4,3], \"otherRecords\":[1,2, 4], \"another\":\"record\"}";