+++ src/test/java/io/json/compare/matcher/diffs/JsonPathDiffTests.java
@@ -56,4 +56,51 @@ public class JsonPathDiffTests {
         String actual2 = "[\"b\",false,{\"x\":[1,2,5]}, {\"w\":\"yyyy\"}]";
         JSONCompare.assertMatches(expected2, actual2);
     }
+
+    @Test
+    public void checkDoNotMatchAnyAndJsonPathsFromArray() {
+        String expected = "[false, {\"#($.length())\":1}, \"b\",{\"x\":{\"#($.length())\":2}}, \"!.*\"]";
+        String actual = "[\"b\",false,{\"x\":[1,2,5]}, {\"w\":\"yyyy\"}]";
+        AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 3 DIFFERENCE.*" +
+                "\\QJson path '$.length()' -> Expected json path result\\E.*" +
+                "1.*But got.*4.*" +
+                "Expected element from position 4 was NOT FOUND:.*" +
+                "Expected condition \"\\Q!.*\\E\" from position 5 was not met. Actual JSON ARRAY has extra elements.*"));
+        JSONCompare.assertNotMatches(expected, actual);
+
+        String expected1 = "[false, {\"#($.length())\":3}, \"b\",{\"x\":{\"#($.length())\":3}}, \"!.*\"]";
+        String actual1 = "[\"b\",false,{\"x\":[1,2,5]}]";
+        JSONCompare.assertMatches(expected1, actual1);
+    }
+
+    @Test
+    public void checkDoNotMatchAnyAndJsonPathsFromObject() {
+        String expected = "{\"#($.length())\":2, \"b\":\"val1\", \"!.*\":\".*\"}";
+        String actual = "{\"b\":\"val1\", \"a\":\"val2\"}";
+        AssertionError error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 1 DIFFERENCE.*" +
+                "Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*"));
+        JSONCompare.assertNotMatches(expected, actual);
+
+
+        String expected1 = "{\"#($.length())\":2, \"b\":\"val1\", \"!name\":\".*\", \"!.*\":\".*\"}";
+        String actual1 = "{\"b\":\"val1\", \"a\":\"val2\"}";
+        error = assertThrows(AssertionError.class, () -> JSONCompare.assertMatches(expected, actual));
+        assertTrue(error.getMessage().matches("(?s).*FOUND 1 DIFFERENCE.*" +
+                "Expected condition '\\Q!.*\\E' was not met. Actual JSON OBJECT has extra fields.*"));
+        JSONCompare.assertNotMatches(expected1, actual1);
+
+        expected1 = "{\"#($.length())\":1, \"b\":\"val1\", \"!name\":\".*\", \"!.*\":\".*\"}";
+        actual1 = "{\"b\":\"val1\"}";
+        JSONCompare.assertMatches(expected1, actual1);
+
+        expected1 = "{\"#($.length())\":2,  \"a\":\"val2\", \"b\":\"val1\", \"!name\":\".*\", \"!.*\":\".*\"}";
+        actual1 = "{\"b\":\"val1\", \"a\":\"val2\"}";
+        JSONCompare.assertMatches(expected1, actual1);
+
+        expected1 = "{\"#($.length())\":0, \"!name\":\".*\", \"!.*\":\".*\"}";
+        actual1 = "{}";
+        JSONCompare.assertMatches(expected1, actual1);
+    }
 }